requirements.txt
argh==0.26.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
dramatiq==1.10.0
fastapi==0.63.0
gevent==21.1.2
greenlet==1.0.0
h11==0.12.0
idna==2.10
pathtools==0.1.2
prometheus-client==0.9.0
pycparser==2.20
pydantic==1.7.3
pyTelegramBotAPI==3.7.6
python-dotenv==0.15.0
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
starlette==0.13.6
urllib3==1.26.2
uvicorn==0.13.3
watchdog==0.8.3
watchdog-gevent==0.1.0
zope.event==4.5.0
zope.interface==5.2.0
